refactor moved expense inside group at internal/ level
it feels right: an expense does not live without a group. Plus, when an expense is added the group balance is modified. That's responsability of the gorup service. At REST api level, it feels correct to keep a separate endpoint, but that's fine, internal should not be affected by outside interfaces. I hope this makes sense